reorganize syllabus components
