+++ frontend/vue/components/Syllabus/SyllabusCreateLayout.vue
