+++ frontend/vue/components/Syllabus/SyllabusLayout.vue
