+++ server/templates/syllabusCreate.pug
@@ -13,7 +13,6 @@ html(lang=lang dir=dir)
       q-syllabus-create-layout
 
     script(type="application/json")#translations!= translationsJSON
-    script(type="application/json")#privacyPolicy!= privacyPolicyMD
 
     script.
       var textbookAnalytics = {
